Changes to be committed: 	modified:   include/tools.c 	new file:   one_process_covert_channel/.clflushtest.c.swp 	modified:   one_process_covert_channel/clflushtest.c 	deleted:    one_process_covert_channel/result.txt 	modified:   one_process_covert_channel/result/clflushtest 	modified:   tags

+++ side_channel/include/tools.c
@@ -1,7 +1,7 @@
 #include "tools.h"
 
 
-void binary_print(size_t addr,int size)
+void binary_print(uint64_t addr,int size)
 {
 	int i;
 	for(i=1;i<=size;i++)
